Moved filter_models and load_results to leaderboard/utils
